persist my sessions, added hooks
